+++ app/src/main/java/net/nashlegend/sourcewall/activities/PostActivity.java
@@ -4,6 +4,7 @@ import android.animation.Animator;
 import android.animation.AnimatorSet;
 import android.animation.ObjectAnimator;
 import android.app.AlertDialog;
+import android.app.Dialog;
 import android.content.BroadcastReceiver;
 import android.content.ClipData;
 import android.content.ClipboardManager;
@@ -30,6 +31,7 @@ import com.umeng.analytics.MobclickAgent;
 import net.nashlegend.sourcewall.R;
 import net.nashlegend.sourcewall.adapters.PostDetailAdapter;
 import net.nashlegend.sourcewall.dialogs.FavorDialog;
+import net.nashlegend.sourcewall.dialogs.ReportDialog;
 import net.nashlegend.sourcewall.model.Post;
 import net.nashlegend.sourcewall.model.UComment;
 import net.nashlegend.sourcewall.request.RequestObject.SimpleCallBack;
@@ -44,13 +46,13 @@ import net.nashlegend.sourcewall.util.Consts;
 import net.nashlegend.sourcewall.util.Mob;
 import net.nashlegend.sourcewall.util.RegUtil;
 import net.nashlegend.sourcewall.util.ShareUtil;
+import net.nashlegend.sourcewall.util.ToastUtil;
 import net.nashlegend.sourcewall.util.UiUtil;
 import net.nashlegend.sourcewall.util.UrlCheckUtil;
 import net.nashlegend.sourcewall.view.MediumListItemView;
 import net.nashlegend.sourcewall.view.common.LoadingView;
 import net.nashlegend.sourcewall.view.common.listview.LListView;
 
-import java.lang.reflect.Method;
 import java.util.ArrayList;
 
 import rx.Observer;
@@ -181,6 +183,9 @@ public class PostActivity extends BaseActivity implements LListView.OnRefreshLis
         switch (id) {
             case android.R.id.home:
                 finish();
+                break;
+            case R.id.action_report:
+                reportPost();
                 break;
             case R.id.action_load_acs:
                 startLoadAcs();
@@ -384,6 +389,65 @@ public class PostActivity extends BaseActivity implements LListView.OnRefreshLis
         });
     }
 
+    private void reportComment(final UComment uComment) {
+        if (!UserAPI.isLoggedIn()) {
+            gotoLogin();
+            return;
+        }
+        new ReportDialog.Builder(this)
+                .setTitle("举报")
+                .setReasonListener(new ReportDialog.ReportReasonListener() {
+                    @Override
+                    public void onGetReason(final Dialog dia, String reason) {
+                        PostAPI.reportReply(uComment.getID(), reason, new SimpleCallBack<Boolean>() {
+                            @Override
+                            public void onFailure() {
+                                ToastUtil.toastBigSingleton("举报未遂……");
+                            }
+
+                            @Override
+                            public void onSuccess() {
+                                UiUtil.dismissDialog(dia);
+                                ToastUtil.toastBigSingleton("举报成功");
+                            }
+                        });
+                    }
+                })
+                .create()
+                .show();
+    }
+
+    private void reportPost() {
+        if (!UserAPI.isLoggedIn()) {
+            gotoLogin();
+            return;
+        }
+        if (post == null) {
+            return;
+        }
+        new ReportDialog.Builder(this)
+                .setTitle("举报")
+                .setReasonListener(new ReportDialog.ReportReasonListener() {
+                    @Override
+                    public void onGetReason(final Dialog dia, String reason) {
+                        PostAPI.reportPost(post.getId(), reason, new SimpleCallBack<Boolean>() {
+                            @Override
+                            public void onFailure() {
+                                ToastUtil.toastBigSingleton("举报未遂……");
+                            }
+
+                            @Override
+                            public void onSuccess() {
+                                UiUtil.dismissDialog(dia);
+                                ToastUtil.toastBigSingleton("举报成功");
+                            }
+                        });
+                    }
+                })
+                .create()
+                .show();
+    }
+
     private void deleteComment(final UComment comment) {
         if (!UserAPI.isLoggedIn()) {
             gotoLogin();
@@ -415,11 +479,14 @@ public class PostActivity extends BaseActivity implements LListView.OnRefreshLis
     private void onReplyItemClick(final View view, int position, long id) {
         if (view instanceof MediumListItemView) {
             final UComment comment = ((MediumListItemView) view).getData();
-            ArrayList<String> ops = new ArrayList<>();
+            final ArrayList<String> ops = new ArrayList<>();
             ops.add(getString(R.string.action_reply));
             ops.add(getString(R.string.action_copy));
             if (!comment.isHasLiked()) {
                 ops.add(getString(R.string.action_like));
+            }
+            if (!comment.getAuthor().getId().equals(UserAPI.getUserID())) {
+                ops.add(getString(R.string.report));
             }
             if (comment.getAuthor().getId().equals(UserAPI.getUserID())) {
                 ops.add(getString(R.string.action_delete));
@@ -428,25 +495,29 @@ public class PostActivity extends BaseActivity implements LListView.OnRefreshLis
             for (int i = 0; i < ops.size(); i++) {
                 operations[i] = ops.get(i);
             }
-            new AlertDialog.Builder(this).setItems(operations, new DialogInterface.OnClickListener() {
-                @Override
-                public void onClick(DialogInterface dialog, int which) {
-                    switch (which) {
-                        case 0:
-                            replyComment(comment);
-                            break;
-                        case 1:
-                            copyComment(comment);
-                            break;
-                        case 2:
-                            likeComment((MediumListItemView) view);
-                            break;
-                        case 3:
-                            deleteComment(comment);
-                            break;
-                    }
-                }
-            }).create().show();
+            new AlertDialog.Builder(this)
+                    .setItems(operations, new DialogInterface.OnClickListener() {
+                        @Override
+                        public void onClick(DialogInterface dialog, int which) {
+                            if (which >= ops.size() || which < 0) {
+                                return;
+                            }
+                            String desc = ops.get(which);
+                            if (desc.equals(getString(R.string.action_reply))) {
+                                replyComment(comment);
+                            } else if (desc.equals(getString(R.string.action_copy))) {
+                                copyComment(comment);
+                            } else if (desc.equals(getString(R.string.action_like))) {
+                                likeComment((MediumListItemView) view);
+                            } else if (desc.equals(getString(R.string.action_delete))) {
+                                deleteComment(comment);
+                            } else if (desc.equals(getString(R.string.report))) {
+                                reportComment(comment);
+                            }
+                        }
+                    })
+                    .create()
+                    .show();
         }
     }
 
@@ -540,7 +611,7 @@ public class PostActivity extends BaseActivity implements LListView.OnRefreshLis
 
                     @Override
                     public void call(ResponseObject<Post> result) {
-                        if (isFinishing()){
+                        if (isFinishing()) {
                             return;
                         }
                         if (result.ok) {
@@ -602,7 +673,7 @@ public class PostActivity extends BaseActivity implements LListView.OnRefreshLis
 
                     @Override
                     public void onNext(ResponseObject<ArrayList<UComment>> result) {
-                        if (isFinishing()){
+                        if (isFinishing()) {
                             return;
                         }
                         if (result.ok) {
